msm: pcie: fix PCIe RAM issue on APQ8064v1
PCIe packet corruption has been observed on APQ8064 CDP, where
bits are flipped when sending packets over a PCIe GigE interface.
The RAM used for transfers was determined to be the source of
the problem. A software solution here is to limit the TLP size from
the default 4KB to 2KB based on the PCIe core version.

Change-Id: Iafda58963643424f7c0d647c487ffe6316037e66
Signed-off-by: Yan He <yanhe@codeaurora.org>